opt/apps/server-a/code/gnh-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
pymongo>=4.6,<5.0
gspread>=6.0,<7.0
google-auth>=2.29,<3.0
PyJWT>=2.8,<3.0
gunicorn>=21.2,<22.0
requests>=2.31,<3.0
